requirements.txt
asgiref==3.8.1
astroid==3.3.5
asttokens==2.4.1
black==24.10.0
boto3==1.35.54
botocore==1.35.54
cfgv==3.4.0
click==8.1.7
decorator==5.1.1
dill==0.3.9
distlib==0.3.9
Django==5.1.2
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.2
executing==2.1.0
filelock==3.16.1
gunicorn==23.0.0
identify==2.6.1
ipython==8.29.0
isort==5.13.2
jedi==0.19.1
jmespath==1.0.1
kafka-python==2.0.2
libcst==1.5.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycln==2.4.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
rich==13.9.4
s3transfer==0.10.3
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomlkit==0.13.2
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
